mock data as fixture - core
